refactor : using MVVM pattern
