--- app/src/main/java/campus/tech/kakao/map/MainActivity.kt
@@ -1,14 +1,0 @@
-package campus.tech.kakao.map
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        val dbHelper = PlaceDBHelper(this)
-//        dbHelper.insertDummyData()
-        dbHelper.readData()
-    }
-}
