+++ app/src/main/java/campus/tech/kakao/map/db/PlaceContract.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.db
 
 import android.provider.BaseColumns
