+++ app/src/main/java/campus/tech/kakao/map/db/PlaceDBHelper.kt
@@ -1,11 +1,10 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.db
 
 import android.content.ContentValues
 import android.content.Context
+import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
-import android.util.Log
-
 
 
 class PlaceDBHelper(context:Context) : SQLiteOpenHelper(context, "place.db", null, 1){
@@ -43,30 +42,9 @@ class PlaceDBHelper(context:Context) : SQLiteOpenHelper(context, "place.db", nul
         db.insert(PlaceContract.PlaceEntry.TABLE_NAME, null, values)
     }
 
-    fun readData() :MutableList<Place>{
+    fun readData(): Cursor {
         val rDb = this.readableDatabase
 
-        val cursor = rDb.rawQuery("SELECT * FROM ${PlaceContract.PlaceEntry.TABLE_NAME};", null)
-
-        val result = mutableListOf<Place>()
-
-        while (cursor.moveToNext()) {
-            val place = Place(
-                cursor.getString(
-                    cursor.getColumnIndexOrThrow(PlaceContract.PlaceEntry.COLUMN_NAME)
-                ),
-                cursor.getString(
-                    cursor.getColumnIndexOrThrow(PlaceContract.PlaceEntry.COLUMN_LOCATION)
-                ),
-                cursor.getString(
-                    cursor.getColumnIndexOrThrow(PlaceContract.PlaceEntry.COLUMN_CATEGORY)
-                )
-            )
-            Log.d("readData", "이름 = ${place.name}, 위치 = ${place.location}, 분류 = ${place.category}")
-            result.add(place)
-        }
-
-        cursor.close()
-        return result
+        return rDb.rawQuery("SELECT * FROM ${PlaceContract.PlaceEntry.TABLE_NAME};", null)
     }
 }
