+++ app/src/main/java/campus/tech/kakao/map/model/Place.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 data class Place(
     val name : String,
