+++ app/src/main/java/campus/tech/kakao/map/viewmodel/MainActivityViewModel.kt
@@ -1,0 +1,10 @@
+package campus.tech.kakao.map.viewmodel
+
+import androidx.lifecycle.ViewModel
+import campus.tech.kakao.map.repository.Repository
+
+class MainActivityViewModel (private val repository: Repository) : ViewModel() {
+    fun getPlace(){
+        repository.getAllPlace()
+    }
+}
