+++ app/src/main/java/campus/tech/kakao/map/viewmodel/ViewModelFactory.kt
@@ -1,0 +1,14 @@
+package campus.tech.kakao.map.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import campus.tech.kakao.map.repository.Repository
+
+class ViewModelFactory(private val repository: Repository) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if(modelClass.isAssignableFrom(MainActivityViewModel::class.java)){
+            return MainActivityViewModel(repository) as T
+        }
+        throw IllegalArgumentException("unKnown ViewModel class")
+    }
+}
